chore: Embedded PID Control.

+++ src/RoasterPrefs.h
@@ -1,5 +1,5 @@
 void setupPreferences();
-void setDoubleValue(const char* name, double value);
-double getDoubleValue(const char* name, double defaultValue);
+void setFloatValue(const char* name, float value);
+float getFloatValue(const char* name, float defaultValue);
 void setLongValue(const char* name, long value);
 long getLongValue(const char* name, long defaultValue);
